#29: Completely reworked the database interfaces , finished user objects, added much more.
New database interface that is implemented by mongodatabase to include mongodb operations. this will be super useful for changing into a temp database or switching databases completely by just creating new implementations. new classes for responses from async functions as well as a builder pattern for building update requests with mongo database. might need to abstract that a little more... New package for running typescript files and a new script to run the database test file i was working on. all in all a good update. Next: will need to create some tests for it to be good and handle error/edge cases. also need to reqork the service-api stuff