#44: Fix add user and update relevant test script
Update user also fixed but that one is a bit wonky because right now it just takes an object as a parameter for which to update the data, but this just replaces the data in the database. What we want to do is have different operations for like adding to a list, removing a property, or setting a property. Maybe follow the MongoDB pattern and then go back to the builder pattern? idk figure out a way to easily create updates and pass them into appropriate functions